内部优化
